Build breakpad / crash reporting on Linux 64-bit.
Review URL: http://codereview.chromium.org/165493


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23396 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/app/breakpad_linux.h
@@ -11,9 +11,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 extern void InitCrashReporter();
 
 #if defined(GOOGLE_CHROME_BUILD)
-static const unsigned kMaxActiveURLSize = 1024;
-static const unsigned kGuidSize = 32;  // 128 bits = 32 chars in hex.
-static const unsigned kDistroSize = 128;
+static const size_t kMaxActiveURLSize = 1024;
+static const size_t kGuidSize = 32;  // 128 bits = 32 chars in hex.
+static const size_t kDistroSize = 128;
 
 struct BreakpadInfo {
   const char* filename;
